Fix form submission cancel for unassociated forms
To export SVG and CSV client data, we add the data to JS forms
and post them on the fly. In newer browser versions, for this to
work, we have to append the forms to the DOM.

With this commit the (hidden) forms get added to the DOM, submitted
and removed.