Fix missing multi select options
Fix bug introduced by django 1.11 update.

QuerySets (value lists) can no longer be passed as values (options)
to multi select widgets. When generating the HTML option elements
the value (if not a list) is encapsulated in a list and cast to
string. That is `<QuerySet [1, 2, 3]>` becomes `[u'<QuerySet [1, 2,
3]>']` whereas it should become `[1, 2, 3]`.

See `django.forms.widgets.ChoiceWidget.format_value`

This commit casts the QuerySet to a list before assigning it to
the widget.