Add basic Dockerfile
The added Dockerfile generates a DMARC viewer container image
based on Alpine using uwsgi to serve the Django app and static
files.

The image size can be reduced by uninstalling build dependencies.
However the corresponding code is currently commented out because
of an issue with a runtime dependency for CairoSVG that is
currently not detected.

The setup is based on a blog post from
https://www.caktusgroup.com/blog/2017/03/14/production-ready-dockerfile-your-python-django-app/