Get GeoIP databases using new pattern
Maxmind no longer offer the databases without authentication.  To get
the DB we use the geoipupdater tool.  This is compiled against libc
rather than musl so we've switched to the debian base image.

During this process I tried to upgrade to python 3.7 or 3.8 but that
broke the Django layers, so sticking to 3.6 for now.